refactor(apple): reuse code
