+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "winio-ui-qt",
     "winio-ui-app-kit",
     "winio-ui-ui-kit",
+    "winio-ui-apple-common",
 ]
 
 resolver = "3"
@@ -38,6 +39,7 @@ winio-ui-gtk = { path = "winio-ui-gtk", version = "0.6.0-rc.1" }
 winio-ui-qt = { path = "winio-ui-qt", version = "0.6.0-rc.1" }
 winio-ui-app-kit = { path = "winio-ui-app-kit", version = "0.6.0-rc.1" }
 winio-ui-ui-kit = { path = "winio-ui-ui-kit", version = "0.1.0-rc.1" }
+winio-ui-apple-common = { path = "winio-ui-apple-common", version = "0.1.0-rc.1" }
 
 compio = { version = "0.19.0-rc.2", default-features = false }
 compio-log = "0.1.0"
