+++ winio-ui-app-kit/src/ui/button.rs
@@ -14,10 +14,7 @@ use winio_callback::Callback;
 use winio_handle::AsContainer;
 use winio_primitive::{Point, Size};
 
-use crate::{
-    GlobalRuntime, Result, catch,
-    ui::{Widget, from_nsstring},
-};
+use crate::{GlobalRuntime, Result, Widget, catch, from_nsstring};
 
 #[derive(Debug)]
 pub struct Button {
