+++ winio-ui-app-kit/src/ui/list_box.rs
@@ -19,10 +19,7 @@ use winio_callback::Callback;
 use winio_handle::AsContainer;
 use winio_primitive::{Point, Size};
 
-use crate::{
-    GlobalRuntime, Result, catch,
-    ui::{Widget, from_cgsize},
-};
+use crate::{GlobalRuntime, Result, Widget, catch, from_cgsize};
 
 #[derive(Debug)]
 pub struct ListBox {
