+++ winio-ui-ui-kit/src/lib.rs
@@ -3,84 +3,14 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![cfg(target_os = "ios")]
 
-use std::panic::AssertUnwindSafe;
-
 #[cfg(feature = "compio-compat")]
 pub use compio::compat::FuturesAdapter as CompioAdapter;
-use objc2::{exception::Exception, rc::Retained};
-use objc2_foundation::NSError;
 pub(crate) use winio_pollable::GlobalRuntime;
+pub(crate) use winio_ui_apple_common::*;
+pub use winio_ui_apple_common::{Error, Result};
 
 mod runtime;
 pub use runtime::*;
 
 mod ui;
 pub use ui::*;
-
-/// Error type for UIKit.
-#[derive(Debug, thiserror::Error)]
-#[non_exhaustive]
-pub enum Error {
-    /// IO error.
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-    /// Objective-C exception.
-    #[error("Objective-C exception: {0:?}")]
-    ObjC(Option<Retained<Exception>>),
-    /// NSError.
-    #[error("NSError: {0:?}")]
-    NS(Option<Retained<NSError>>),
-    /// Channel recv error.
-    #[error("Channel recv error: {0}")]
-    ChannelRecv(#[from] local_sync::oneshot::error::RecvError),
-    /// Null pointer returned.
-    #[error("Null pointer returned")]
-    NullPointer,
-    /// Called from non-main thread.
-    #[error("Called from non-main thread")]
-    NotMainThread,
-    /// Feature not supported.
-    #[error("Feature not supported")]
-    NotSupported,
-}
-
-// SAFETY: NSException & NSError are thread-safe.
-unsafe impl Send for Error {}
-unsafe impl Sync for Error {}
-
-impl From<Retained<Exception>> for Error {
-    fn from(exc: Retained<Exception>) -> Self {
-        Error::ObjC(Some(exc))
-    }
-}
-
-impl From<Option<Retained<Exception>>> for Error {
-    fn from(exc: Option<Retained<Exception>>) -> Self {
-        Error::ObjC(exc)
-    }
-}
-
-impl From<Retained<NSError>> for Error {
-    fn from(err: Retained<NSError>) -> Self {
-        Error::NS(Some(err))
-    }
-}
-
-impl From<Option<Retained<NSError>>> for Error {
-    fn from(err: Option<Retained<NSError>>) -> Self {
-        Error::NS(err)
-    }
-}
-
-/// Result type for UIKit.
-pub type Result<T, E = Error> = std::result::Result<T, E>;
-
-pub(crate) fn catch<F, R>(f: F) -> Result<R>
-where
-    F: FnOnce() -> R,
-{
-    match objc2::exception::catch(AssertUnwindSafe(f)) {
-        Ok(v) => Ok(v),
-        Err(exc) => Err(Error::from(exc)),
-    }
-}
