+++ winio-ui-ui-kit/src/ui/label.rs
@@ -55,7 +55,7 @@ impl Label {
     pub fn set_tooltip(&mut self, s: impl AsRef<str>) -> Result<()>;
 
     pub fn text(&self) -> Result<String> {
-        catch(|| crate::ui::from_nsstring(&self.view.text().unwrap_or_default()))
+        catch(|| crate::from_nsstring(&self.view.text().unwrap_or_default()))
     }
 
     pub fn set_text(&mut self, s: impl AsRef<str>) -> Result<()> {
